requirements/local.txt
-r base.txt

pytest==7.4.0
pytest-django==4.5.2

factory-boy==3.2.1
Faker==18.6.0
model-bakery==1.17.0

ipdb==0.13.13
ipython==8.10.0

django-debug-toolbar==3.8.1

mypy==1.2.0

django-stubs==1.14.0
djangorestframework-stubs==1.8.0
boto3-stubs==1.26.81

flake8==6.0.0
isort==5.12.0
black==23.3.0
pre-commit==3.2.2
